# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dushi Stores Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "info@dushistores.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get All Products",
                "description": "get all products",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/show/{product_id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Product Details",
                "description": "get product detail",
                "operationId": "03a3d11d2dd72a2c96cc4622572fbaad",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/category/{category_id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Products By Category",
                "description": "get products by category",
                "operationId": "36f8a4e75f02a0726a8a5e41b5b56fca",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/subcategory/{sub_category_id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Products By SubCategory",
                "description": "get products by subcategory",
                "operationId": "458db33ac05430473de2c1de9682bf53",
                "parameters": [
                    {
                        "name": "sub_category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/deals": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Disconted Products",
                "description": "get discounted products",
                "operationId": "d1a823a732edd4b21d52679f6581ef8c",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/top": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Top Selling Products",
                "description": "get top selling products",
                "operationId": "950dea6509757b621b60df9034c2a288",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/new": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Products By New Arrivals",
                "description": "get products by new arrivals",
                "operationId": "d96d9b083ebd11c39f9041815ac7c796",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/filter": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Filter Products By (price range)",
                "description": "get discounted products",
                "operationId": "dc537a7dfbacebd44641b056586b4ef9",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/sort": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Sort Products By (name, sale, price)",
                "description": "get discounted products",
                "operationId": "c77a85534522d6f64c85faebb7312442",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "sale",
                                "price"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/products/search/{search}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Search Products",
                "description": "search products",
                "operationId": "c419dec97880e2c5ec7c74936b02ac69",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}